Update: refactor BookingModal with new components for date, time, guest selection, and confirmation
